Group all legacy checkout tests inside a subfolder
