+++ tests/e2e/tests/_legacy-experience/checkout/sca-card.spec.js
@@ -1,24 +1,19 @@
 import { test, expect } from '@playwright/test';
 import config from 'config';
-import { payments } from '../../utils';
+import { payments } from '../../../utils';
 
 const {
 	emptyCart,
 	setupProductCheckout,
-	setupBlocksCheckout,
+	setupCheckout,
 	fillCardDetails,
 } = payments;
 
-test( 'customer can checkout with a SCA card @smoke @blocks', async ( {
-	page,
-} ) => {
+test( 'customer can checkout with a SCA card @smoke', async ( { page } ) => {
 	await emptyCart( page );
 
 	await setupProductCheckout( page );
-	await setupBlocksCheckout(
-		page,
-		config.get( 'addresses.customer.billing' )
-	);
+	await setupCheckout( page, config.get( 'addresses.customer.billing' ) );
 	await fillCardDetails( page, config.get( 'cards.3ds' ) );
 	await page.locator( 'text=Place order' ).click();
