+++ tests/e2e/tests/_legacy-experience/checkout/subscription-product.spec.js
@@ -1,8 +1,8 @@
 import { test, expect } from '@playwright/test';
 import config from 'config';
-import { payments, api } from '../../utils';
+import { payments, api } from '../../../utils';
 
-const { setupBlocksCheckout, fillCardDetails } = payments;
+const { setupCheckout, fillCardDetails } = payments;
 
 let productId;
 
@@ -29,7 +29,7 @@ test.afterAll( async () => {
 	await api.deletePost.product( productId );
 } );
 
-test( 'customer can purchase a subscription product @smoke @blocks @subscriptions', async ( {
+test( 'customer can purchase a subscription product @smoke @subscriptions', async ( {
 	page,
 } ) => {
 	await page.goto( `?p=${ productId }` );
@@ -42,7 +42,7 @@ test( 'customer can purchase a subscription product @smoke @blocks @subscription
 			Date.now() + '+' + config.get( 'addresses.customer.billing.email' ),
 	};
 
-	await setupBlocksCheckout( page, customerData );
+	await setupCheckout( page, customerData );
 	await fillCardDetails( page, config.get( 'cards.basic' ) );
 
 	await page.locator( 'text=Sign up now' ).click();
