+++ tests/e2e/tests/_legacy-experience/subscriptions/subscription-renewal.spec.js
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import config from 'config';
-import { payments, api, user } from '../../utils';
+import { payments, api, user } from '../../../utils';
 
 const { setupCheckout, fillCardDetails } = payments;
