+++ tests/e2e/tests/_legacy-experience/woocommerce-blocks/normal-card.spec.js
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import config from 'config';
-import { payments } from '../../utils';
+import { payments } from '../../../utils';
 
 const {
 	emptyCart,
